Добавил немного связанных фреймворков, без которых проект не собирался

+++ Moto64/Moto64/MTRSSDataSource.m
@@ -57,7 +57,7 @@
     
     NSError *error = nil;
     
-    int count = [self.managedObjectContext countForFetchRequest:fetchRequest error:&error];
+    NSUInteger count = [self.managedObjectContext countForFetchRequest:fetchRequest error:&error];
     
     return count > 0;
 }
